dom/nodes/insertion-effects/Node-appendChild-script-and-stylesheet-link.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting script and stylesheet link from a div</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<script>
const happened = [];
const link = document.createElement("link");
test(() => {
  link.rel = "stylesheet";
  link.href = "data:text/css,";

  const script = document.createElement("script");
  script.textContent = `
    happened.push(link.sheet === null);
`;

  const div = document.createElement("div");
  div.appendChild(script);
  div.appendChild(link);

  assert_array_equals(happened, []);
  document.body.appendChild(div);
  assert_array_equals(happened, [true]);
});
</script>
